feat : Restart

+++ Assets/01.Code/Entities/IDamageable.cs
@@ -2,6 +2,6 @@
 {
     public interface IDamageable
     {
-        public void ApplyDamage(int damage, Entity dealer);
+        public void ApplyDamage(float damage, Entity dealer);
     }
 }
